refactor: extract features from reviews components (#509)
