+++ apps/container/src/features/reviews/components/PendingReviewEnrollmentList/PendingReviewEnrollmentList.vue
@@ -26,9 +26,10 @@ import { useQuery } from '@tanstack/vue-query';
 import { Enrollment,Enrollments  } from '@ufabc-next/services';
 import { computed } from 'vue';
 
-import { PendingReviewEnrollment } from '@/components/PendingReviewEnrollment';
 import { FeedbackAlert } from '@/components/ui/FeedbackAlert';
 import { PaperCard } from '@/components/ui/PaperCard';
+
+import { PendingReviewEnrollment } from '../PendingReviewEnrollment';
 
 const { data: enrollments, isError: isErrorEnrollment } = useQuery({
   refetchOnWindowFocus: false,
